Version 1.0
This is the first stone put into place of a continuous project. It will be my life's goal in being able to tell this story. This game is a text adventure or interactive fiction depending on what you prefer. It's written in Qbasic to stick with the original game I made years ago. In this version, you have the standard enemies and items scattered about the castle. There are two passages within the game so far that aren't connected to the other sections of the map. Those will be added next time in a future update. This game is always going to be expanding and evolving until it's a fully functional world of enjoyment and excitement.