Version 1.45 Update
This is a quick update in which I updated and corrected all errors with the directional mapping system. Which moves the player between rooms. Within this update, I have added a new section to which is called "Temple Garden". With it, I added a few more NPCs with which you can converse and several objects that you can examine and picture in your mind.