### v1.6 Whats New
Welcome to the great library. A place to which one can venture forth and learn the great history of the lands that surround you. The very idea of discovering such a place to me is one of immense joy.
I've always been a fan of knowledge and being able to discover such a place where one can read. I love the idea of the player wandering through this immense library filled with books that you can read. Not also the fun of reading about the history of the land and the lands that surround the kingdom.
I've taken it upon myself to add many ghosts and lost souls to this section of the game. I always loved the idea of ghosts and lost souls wandering about the grounds of the library. It's something of a fun means of interacting with the souls that have yet to leave this plain of existence. You can speak with several of these wonderful characters who are spread out in the library going about their business and attempting to help you on your journey.
It's one of the many aspects of this world. Reading up on those foul creatures that lurk in the darkness and wait to feast on your flesh and gnaw on your bones. It also gives the player a greater understanding of the world around them that is separated from the world we all live in. I am also adding in a vast world map that you can't take with you on your journey but can examine and map out on your own. Though, I'm divided about taking it with you on your journey or just leaving it there for the next adventurer to use.
This in my opinion is the vast hub to which you can learn and plot your next course of action. Plus the very thought of being able to meet with the Librarian in control of this entire place is nothing short of fun in my opinion. Between him and yourself, you are the only living wanderer within that place.
I always pictured such wonderful games that have a library in them. To me, it's a dying room to have a player explore, and plus you can do so much and add so many traps, monsters, and funny little nods to other great works of fiction. Plus, I always loved the very thought of exploring the last reminder of any dying society's existence.
The layout of the library is that of a 3x3 design consisting of 9 moves per section. Each move is connected by a single section that connects to another 3x3 design consisting of other portions of this great realm. There are nine sections in total and a hidden section consisting of a section that's 2x3. This is the place where the map resides and just for a note that there is no exit from this place other than the way you came in. With the great library comes a sense of wonder for me. It's a place where you can explore the world from the comfort of your local library.

With the great library completed and with the next section which consists of the Witch who knows about the vast adventure that you the player have allowed yourself to become a part of.
The map itself starts at room 614 and goes about that of a path that takes you to discover a few more twists and turns. The player is greeted with a sign warning them to stay away but obviously, the very thought of staying away from such a place is crazy in my opinion. But I figured I will give the player the option. Now, with this game, you don't need to explore every inch of the world but I hope you do, I believe in giving the player options, options to find out the secrets of the land, meeting with the characters, and getting a feel for this place.