Version 1.8.1
In this update, I added the village of Frun which takes place within a place on the brink of death. I've added 13 new NPCs and 6 interactive objects within the game. I also rearranged some of the code and edited some of the interactive objects that I decided to edit out.